repair pemeriksaan kesehatan

+++ Bintangku.WebApi/Data/DTO/Pemeriksaan/PemeriksaanStatusGiziImtUDto.cs
@@ -2,8 +2,8 @@ namespace Bintangku.WebApi.Data.DTO.Pemeriksaan
 {
     public class PemeriksaanStatusGiziImtUDto
     {
-        public int BeratBadan { get; set; }
-        public int TinggiBadan { get; set; }
+        public double BeratBadan { get; set; }
+        public double TinggiBadan { get; set; }
         public int Umur { get; set; }
         public int ZCode { get; set; } 
     }
