+++ Bintangku.WebApi/Data/Entities/Pemeriksaan/PemeriksaanStatusGiziImtU.cs
@@ -3,10 +3,10 @@ namespace Bintangku.WebApi.Data.Entities.Pemeriksaan
     public class PemeriksaanStatusGiziImtU
     {
         public int PemeriksaanStatusGiziImtUId { get; set; } 
-        public int BeratBadan { get; set; }
-        public int TinggiBadan { get; set; }
+        public double BeratBadan { get; set; }
+        public double TinggiBadan { get; set; }
         public int Umur { get; set; }
-        public int IMT { get; set; }
+        public double IMT { get; set; }
         public int ZCode { get; set; }
         public string StatusGizi { get; set; }
         public string Tindakan { get; set; }
